requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pydantic
faiss-cpu
langchain
sentence-transformers
google-auth-oauthlib
google-api-python-client
streamlit